vclib.git: Restructure modules, per backend module (pygit2 and no backend)
